Extract useSmartDefaultFilter
